refactor: break out syntax_highlight mod into a file
